Prevent the database file from being re-written.
Remove the functionality to automatically create the database schema if
the database is empty. This is to help find out when the database is set
to NULL. Also echo when games is set to NULL to help debug.

XXX: This commit is a temporary fix and aims to help debugging.